// apps/web/src/app/sub/masters/types.ts
export type PlayerMarker = "none" | "successor" | "not-qualified";

export interface PlayerRowData {
  player: Player;
  index: number;
  marker: PlayerMarker;
}

export interface Player {
  id: string;
	name: string;
	tournamentPoints: number;
	buchholz: number;
  rankSum: number;
  averageRank: number;
  playedTournaments: number;
  [key: `tournament${number}`]: number | undefined;
	// hasPlayedThreeOrMoreTournaments?: boolean;
}

export interface TournamentResultPlayer {
  "rank": number;
  "playerId": string;
  "name": string;
  "rating": number;
  "score": number;
  "buchholz": number;
}

export interface TournamentResult {
  "data": {
    "tournamentNumber": number;
    "tournamentName": string;
    "totalRounds": number;
    "pointRule": "1-point" | "3-point";
    "rows": TournamentResultPlayer[];
  }
}

export interface TournamentDataTypes {
	name: string;
	status: "completed" | "upcoming";
	value: string;
	date: string | null;
	startTime: string | null;
	endTime: string | null;
	type: "blitz" | "rapid" | "960" | "final";
	mode: string | null;
	pointRule: "1-point" | "3-point" | null;
	evaluation: string | null;
	entryFee: {
		preRegistration: string;
		normalRegistration: string;
	} | null;
	venue: string | null;
	announcementPath: string | null;
	listOfAttendeesUrl: string | null;
	generalInformation: string | null;
}